+++ toy/src/vertex_mesh.cpp
@@ -12,17 +12,25 @@ VertexMesh::~VertexMesh()
 }
 
 void VertexMesh::init_resource(VulkanContext& ctx, std::vector<ColorTextureTriangleData> data)
-{
+{	
+	if (data.empty()) {
+		return;
+	}
+
 	m_ctx = &ctx;
 	// vertex buffer
 	auto vertex_buffer_and_memory = ctx.create_vertex_buffer((uint8_t*)data.data(), stlutil::vector_buffer_size(data));
 	m_vertex_buffer.buffer = vertex_buffer_and_memory.first;
 	m_vertex_buffer.memory = vertex_buffer_and_memory.second;
 	m_vertex_count = data.size() * 3;
+	m_valid = true;
 }
 
 void VertexMesh::draw(VkCommandBuffer command_buffer)
 {
+	if (!m_valid) {
+		return;
+	}
 	VkDeviceSize offset = 0;
 	vkCmdBindVertexBuffers(command_buffer, 0, 1, &m_vertex_buffer.buffer, &offset);
 
@@ -31,6 +39,9 @@ void VertexMesh::draw(VkCommandBuffer command_buffer)
 
 void VertexMesh::destroy()
 {
+	if (!m_valid) {
+		return;
+	}
 	VulkanContext& ctx = *m_ctx;
 	ctx.destroy_vulkan_buffer(m_vertex_buffer);
 }
